# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cartoon-videos
    description: Legally sourced cartoon video catalog
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cartoon-videos:
    get:
      operationId: listCartoonVideos
      tags: [cartoon-videos]
      summary: List cartoon videos
      description: Returns the public, read-only catalog of licensed cartoon videos
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Optional exact category filter
      responses:
        "200":
          description: Cartoon video catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  videos:
                    type: array
                    items:
                      $ref: "#/components/schemas/CartoonVideo"
                required:
                  - videos
  /cartoon-videos/{videoId}:
    get:
      operationId: getCartoonVideo
      tags: [cartoon-videos]
      summary: Get a cartoon video
      description: Returns one licensed cartoon video by its stable ID
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Cartoon video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartoonVideo"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CartoonVideo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        thumbnailUrl:
          type: string
        videoUrl:
          type: string
        description:
          type: string
        category:
          type: string
        durationSeconds:
          type: number
          minimum: 1
        license:
          type: string
        sourceUrl:
          type: string
      required:
        - id
        - title
        - thumbnailUrl
        - videoUrl
        - description
        - category
        - durationSeconds
        - license
        - sourceUrl
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error