# lib/api-zod/src/generated/types/cartoonVideo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartoonVideo {
  id: string;
  title: string;
  thumbnailUrl: string;
  videoUrl: string;
  description: string;
  category: string;
  /** @minimum 1 */
  durationSeconds: number;
  license: string;
  sourceUrl: string;
}